Change persona selection to search-by-username
The persona ID, in the end, wasn't enough to identify a persona in the
API. The API also needs the ID of the platform (pc/xbox/ps4, etc.).
Instead of having to maintain another input, this commit overhauls the
URL structure and layout so that the username, persona ID and platform
int is taken from the URL and produced via the react-select search bar.