Make dark mode switch happen immediately
Previously, dark mode / light mode switch happened by setting the dark
mode cookie, then reloading. This was a quick hack.

Now, we want to be able to invert the attachment image in dark mode, so
we need to access that information via context. This commit extracts the
dark mode functionality to AppContext, and applies it on the frontend on
the fly.